Move the Base64Provider tests in the existing test class
